% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{ALSA_app}
\alias{ALSA_app}
\title{Agricultural Land Suitability Analysis (ALSA) Shiny App}
\usage{
ALSA_app(...)
}
\arguments{
\item{...}{Additional arguments to pass to the Shiny app.}
}
\description{
This function creates a Shiny application for conducting agricultural land suitability analysis.
}
