% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_suitability_potential_table.R
\name{calculate_suitability_potential_table}
\alias{calculate_suitability_potential_table}
\title{Calculate Suitability Potential Table}
\usage{
calculate_suitability_potential_table(
  lookup_intervention,
  intervention_level,
  suitability_attr,
  lookup_suitability_layer
)
}
\arguments{
\item{lookup_intervention}{A tibble containing lookup intervention data.}

\item{intervention_level}{A string indicating the level of intervention.}

\item{suitability_attr}{A dataframe containing suitability attributes with columns name_parameter, ID, intervention, limiting_factor_actual, and suitability.}

\item{lookup_suitability_layer}{A dataframe with two columns: \code{ID} and \code{names}.
\code{ID} is a sequence from 1 to the number of layers in the \code{SpatRaster} object,
and \code{names} are the names of the layers in the \code{SpatRaster} object.}
}
\value{
A tibble containing the calculated suitability potential table.
}
\description{
This function reads a CSV file for lookup intervention data and processes it along with a given suitability attribute dataframe to calculate the suitability potential table.
}
\examples{
\dontrun{
  path <- "data/lookup_tables/lookup_intervention.csv"
  level <- "high"
  attr <- readr::read_csv("output/kesesuaian_jagung_aktual.csv")
  calculate_suitability_potential_table(path, level, attr)
}
}
