% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_and_stack_suitability_factors.R
\name{classify_and_stack_suitability_factors}
\alias{classify_and_stack_suitability_factors}
\title{Classify and Stack Suitability Factors}
\usage{
classify_and_stack_suitability_factors(stacked_raster, suitability_data)
}
\arguments{
\item{stacked_raster}{A stacked \code{SpatRaster} object.}

\item{suitability_data}{A data frame containing crop suitability parameters.}
}
\value{
A stacked \code{SpatRaster} object with classified layers.
}
\description{
This function processes each layer of a stacked \code{SpatRaster} object
using the \code{classify_suitability_predictors} function and returns a stacked
raster with classified layers.
}
\examples{
\dontrun{
  # Assuming you have suitability_factors
  # as a SpatRaster and crop_suitability as a data frame
  stacked_suitability <- classify_and_stack_suitability_factors(
                              suitability_factors, crop_suitability)
}
}
