% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_suitability_predictors.R
\name{classify_suitability_predictors}
\alias{classify_suitability_predictors}
\title{Classify raster based on crop suitability parameters}
\usage{
classify_suitability_predictors(raster_input, suitability_data)
}
\arguments{
\item{raster_input}{A SpatRaster object.}

\item{suitability_data}{A data frame containing crop suitability parameters.}
}
\value{
A reclassified SpatRaster.
}
\description{
Classify raster based on crop suitability parameters
}
\examples{
\dontrun{
classified_raster <- classify_suitability_predictors(
                      raster_input = clim_temperature_avg,
                      suitability_data = crop_suitability)
}
}
