% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_layer_dataframe.R
\name{create_layer_dataframe}
\alias{create_layer_dataframe}
\title{Create Layer Dataframe from SpatRaster}
\usage{
create_layer_dataframe(spatraster)
}
\arguments{
\item{spatraster}{A \code{SpatRaster} object.}
}
\value{
A dataframe with two columns: \code{ID} and \code{names}.
\code{ID} is a sequence from 1 to the number of layers in the \code{SpatRaster} object,
and \code{names} are the names of the layers in the \code{SpatRaster} object.
}
\description{
This function takes a \code{SpatRaster} object and returns a dataframe with two columns:
\code{ID} and \code{names}. \code{ID} is a sequence number starting from 1 for each layer in the
SpatRaster, and \code{names} are the names of these layers.
}
\examples{
\dontrun{
# Assuming you have a SpatRaster object named suitability_factors
# suitability_factors <- ...
dataframe <- create_layer_dataframe(suitability_factors)
print(dataframe)
}

}
