% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_load_crop_suitability_params.R
\name{cropSuitabilityParamsServer}
\alias{cropSuitabilityParamsServer}
\title{Crop Suitability Parameters Module Server}
\usage{
cropSuitabilityParamsServer(id, submittedData)
}
\arguments{
\item{id}{The ID of the module, used to create namespaces for the server elements.}

\item{submittedData}{reactive values to store submitted data}
}
\value{
The server logic for the Crop Suitability Parameters module.
}
\description{
This module server handles the logic for uploading and managing crop suitability parameters.
}
