% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_load_crop_suitability_params.R
\name{cropSuitabilityParamsUI}
\alias{cropSuitabilityParamsUI}
\title{Crop Suitability Parameters Module UI}
\usage{
cropSuitabilityParamsUI(id)
}
\arguments{
\item{id}{The ID of the module, used to create namespaces for the UI elements.}
}
\value{
A tagList containing the UI elements for the Crop Suitability Parameters module.
}
\description{
This module provides a user interface for uploading and managing crop suitability parameters.
}
