% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_load_intervention_options.R
\name{interventionLookupServer}
\alias{interventionLookupServer}
\title{Intervention Lookup Module Server}
\usage{
interventionLookupServer(id, submittedData)
}
\arguments{
\item{id}{The ID of the module, used to create namespaces for the server elements.}

\item{submittedData}{reactive values to store submitted data}
}
\value{
The server logic for the Intervention Lookup module.
}
\description{
This module server handles the logic for uploading and managing intervention lookup tables.
}
