% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_load_intervention_options.R
\name{interventionLookupUI}
\alias{interventionLookupUI}
\title{Intervention Lookup Module UI}
\usage{
interventionLookupUI(id)
}
\arguments{
\item{id}{The ID of the module, used to create namespaces for the UI elements.}
}
\value{
A tagList containing the UI elements for the Intervention Lookup module.
}
\description{
This module provides a user interface for uploading and managing intervention lookup tables.
}
