% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{LaSEM_app}
\alias{LaSEM_app}
\title{Land Suitability Evaluation Module (LaSEM) Shiny App}
\usage{
LaSEM_app(...)
}
\arguments{
\item{...}{Additional arguments to pass to the Shiny app.}
}
\description{
This function creates a Shiny application for conducting agricultural land suitability analysis.
}
