% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_texture_code.R
\name{map_texture_code}
\alias{map_texture_code}
\title{Map Texture Code to TEXTURE_USDA Value}
\usage{
map_texture_code(texture_code, lookup_table)
}
\arguments{
\item{texture_code}{A character string representing the texture code.}

\item{lookup_table}{A data frame containing the mapping between texture codes and TEXTURE_USDA values.}
}
\value{
A character string representing the TEXTURE_USDA value corresponding to the input texture code.
If the texture code is not found in the lookup table, the function returns NA.
}
\description{
This function maps a texture code to its corresponding TEXTURE_USDA value using a lookup table.
}
\examples{
lookup_table <- data.frame(
  texture_kemtan = c("1", "2", "3"),
  TEXTURE_USDA = c("Sandy", "Loamy", "Clayey")
)

map_texture_code("1", lookup_table)  # Returns "Sandy"
map_texture_code("4", lookup_table)  # Returns NA

}
