% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_names_analysed_info.R
\name{print_names_analysed_info}
\alias{print_names_analysed_info}
\title{Print Information about Analyzed Names}
\usage{
print_names_analysed_info(names_analysed)
}
\arguments{
\item{names_analysed}{A vector of names (character) to be analyzed.}
}
\value{
None. This function is used for its side effect of printing
to the console.
}
\description{
This function prints the number of objects and their names
contained in the input vector. It's useful for quickly viewing
the contents and the count of elements in the vector.
}
\examples{
\dontrun{
names_analysed <- c("clim_temperature_avg", "clim_precipitation_tot",
                    "clim_humidity", "soil_coarse_fragments", "soil_depth",
                    "soil_cec", "soil_base_saturation", "soil_ph",
                    "soil_organic_c", "soil_nitrogen", "soil_salinity",
                    "soil_esp", "soil_slope")
print_names_analysed_info(names_analysed)
}

}
