% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_raster_files.R
\name{read_raster_files}
\alias{read_raster_files}
\title{Read Raster Files}
\usage{
read_raster_files(climate_soil_data)
}
\arguments{
\item{climate_soil_data}{A tibble or data frame containing the paths to raster files (\code{raster_path}).}
}
\value{
A list of SpatRaster objects representing the loaded raster files.
}
\description{
This function reads raster files from the provided file paths and returns a list of raster objects.
If a file is not found, it attempts to load the raster from the 'ALSA' package.
}
\examples{
\dontrun{
# Assuming climate_soil_data is predefined
raster_objects <- read_raster_files(climate_soil_data)
}
}
