% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_load_soil_climate_data.R
\name{soilClimateDataServer}
\alias{soilClimateDataServer}
\title{Soil and Climate Data Module Server}
\usage{
soilClimateDataServer(id, submittedData)
}
\arguments{
\item{id}{The ID of the module, used to create namespaces for the server elements.}
}
\value{
The server logic for the Soil and Climate Data module.
}
\description{
This module server handles the logic for uploading, processing, and visualizing soil and climate data.
}
