% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_load_soil_climate_data.R
\name{soilClimateDataUI}
\alias{soilClimateDataUI}
\title{Soil and Climate Data Module UI}
\usage{
soilClimateDataUI(id)
}
\arguments{
\item{id}{The ID of the module, used to create namespaces for the UI elements.}
}
\value{
A tagList containing the UI elements for the Soil and Climate Data module.
}
\description{
This module provides a user interface for uploading and visualizing soil and climate data.
}
