% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack_raster_layers.R
\name{stack_raster_layers}
\alias{stack_raster_layers}
\title{Stack Raster Layers}
\usage{
stack_raster_layers(raster_list, parameter_names)
}
\arguments{
\item{raster_list}{A list of SpatRaster objects to be stacked.}

\item{parameter_names}{A vector of parameter names corresponding to each raster layer in the list.}
}
\value{
A SpatRaster object representing the stacked raster layers with variable names set.
}
\description{
This function stacks a list of raster layers and sets the variable names based on the corresponding parameter names.
}
\examples{
\dontrun{
# Assuming raster_list and parameter_names are predefined
stacked_rasters <- stack_raster_layers(raster_list, parameter_names)
}
}
