% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_suitability_analysis.R
\name{suitabilityAnalysisUI}
\alias{suitabilityAnalysisUI}
\alias{suitabilityAnalysisServer}
\title{Suitability Analysis Module}
\usage{
suitabilityAnalysisUI(id)

suitabilityAnalysisServer(id, submittedData)
}
\arguments{
\item{id}{The module ID.}

\item{submittedData}{reactive values to store submitted data}
}
\description{
This module provides a user interface and server logic for performing suitability analysis
based on soil and climate data, crop parameters, and intervention lookup data. It displays
the suitability map, suitability by factors, suitability polygon data table, and provides
an option to download the suitability analysis results.
}
